+++ scripts/powertrace.py
@@ -7,12 +7,12 @@ import sys, os, sim
 class PowerTrace:
   def setup(self, args):
     args = dict(enumerate((args or '').split(':')))
-    interval_ns = long(args.get(0, '') or 1000000)
+    interval_ns = int(args.get(0, '') or 1000000)
     sim.util.Every(interval_ns * sim.util.Time.NS, self.periodic, roi_only = True)
     self.t_last = 0
 
   def periodic(self, time, time_delta):
-    time = long(time/1e6) # fs to ns
+    time = int(time/1e6) # fs to ns
     if time <= 100:
       # ignore first callback which is at 100ns
       return
@@ -21,7 +21,7 @@ class PowerTrace:
     self.t_last = time
 
   def hook_roi_end(self):
-    self.t_roi_end = long(long(sim.stats.get('performance_model', 0, 'elapsed_time'))/1e6)
+    self.t_roi_end = int(int(sim.stats.get('performance_model', 0, 'elapsed_time'))/1e6)
 
   def hook_sim_end(self):
     self.do_power(self.t_last, None)
